Add files via upload
Experiment currently contains one function, identify, and aims to recursively identify at each level what is a statement and what is an operator. It returns a list of lists that consist of 
-the statements at that level
-the operators at that level
-the level of the evaluation

Basically, brackets get peeled away one level at a time.

Note: This code does not work for all scenarios where it should; in the last else statement, the 'for' loop gets broken by the return statement within it. I don't know how to fix this yet.